+++ ui-tests/vwc-textfield/index.js
@@ -1,3 +1,4 @@
+import vvdCore from '@vonage/vvd-core';
 import { VWCTextField } from '@vonage/vwc-textfield';
 import * as stories from '@vonage/vwc-textfield/stories/textfield-all.stories';
 import { render } from 'lit-html';
@@ -27,9 +28,12 @@ export async function createElementVariations(wrapper) {
 	await new Promise(res => setTimeout(() => {
 		[...textElementWrapper.querySelectorAll('vwc-textfield')].forEach((child) => {
 			child.reportValidity();
+			child.firstElementChild.blur();
 		});
 		res();
 	}, 0));
+
+	await vvdCore.settled;
 }